feature/noteapi

+++ src/Abarnathy.DemographicsAPI/Test/Abarnathy.DemographicsAPI.Test.Integration/Utilities.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using Abarnathy.DemographicsAPI.Data;
 using Abarnathy.DemographicsAPI.Models;
+using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.VisualStudio.TestPlatform.CommunicationUtilities;
 
 namespace Abarnathy.DemographicsAPI.Test.Integration
@@ -9,14 +10,42 @@ namespace Abarnathy.DemographicsAPI.Test.Integration
     {
         public static void InitializeDbForTests(DemographicsDbContext db)
         {
-            db.Patient.AddRange(GetSeedingPatients());
+            var patients = GetSeedingPatients();
+            
+            db.Address.Add(new Address
+            {
+                Id = 1,
+                StreetName = "Baker St",
+                HouseNumber = "6",
+                Town = "Baskerville",
+                State = "Washington",
+                ZipCode = "12345"
+            });
+
+            db.PhoneNumber.Add(new PhoneNumber
+            {
+                Number = "1234567890"
+            });
+            
+            db.Patient.AddRange(patients);
+            
             db.SaveChanges();
         }
 
         public static void ReinitializeDbForTests(DemographicsDbContext db)
         {
-            db.Patient.RemoveRange(db.Patient);
+            WipeDb(db);
             InitializeDbForTests(db);
+        }
+
+        public static void WipeDb(DemographicsDbContext db)
+        {
+            if (db.Patient.Any())
+            {
+                db.Patient.RemoveRange(db.Patient);
+            }
+            
+            db.SaveChanges();
         }
         
         public static IEnumerable<Patient> GetSeedingPatients()
@@ -25,13 +54,17 @@ namespace Abarnathy.DemographicsAPI.Test.Integration
             {
                 new Patient
                 {
+                    Id = 1,
                     FamilyName = "Doe",
                     GivenName = "Jane",
+                    SexId = 2,
                 },
                 new Patient
                 {
+                    Id = 2,
                     FamilyName = "Smith",
                     GivenName = "John",
+                    SexId = 1
                 }
             };
